Lognorm docs improvement and testing and bug fixes

+++ sc/app/objects.py
@@ -5,7 +5,8 @@ ShakeCast to run. These objects are used in the functions.py program
 try:
     import urllib2
 except:
-    import urllib as urllib2
+    import urllib.request as urllib2
+
 import ssl
 import json
 import os
@@ -212,21 +213,23 @@ class ProductGrabber(object):
     def get_event_map(event):
         if not os.path.exists(event.directory_name):
                 os.makedirs(event.directory_name)
-        sc=SC()
-        # download the google maps image
-        url_opener = URLOpener()
-        gmap = url_opener.open("https://api.mapbox.com/styles/v1/mapbox/streets-v10/static/pin-s+F00(%s,%s)/%s,%s,5/200x200?access_token=%s" % (event.lon,
-                                          event.lat,
-                                           event.lon,
-                                           event.lat,
-                                           sc.map_key))
-        
-        # and save it
+
         image_loc = os.path.join(event.directory_name,
                                  'image.png')
-        image = open(image_loc, 'wb')
-        image.write(gmap)
-        image.close()
+
+        if os.path.exists(image_loc) is False:
+            sc=SC()
+            # download the google maps image
+            url_opener = URLOpener()
+            gmap = url_opener.open("https://api.mapbox.com/styles/v1/mapbox/streets-v10/static/pin-s+F00(%s,%s)/%s,%s,5/200x200?access_token=%s" % (event.lon,
+                                            event.lat,
+                                            event.lon,
+                                            event.lat,
+                                            sc.map_key))
+            # and save it
+            image = open(image_loc, 'wb')
+            image.write(gmap)
+            image.close()
             
     def get_new_shakemaps(self, scenario=False):
         """
@@ -273,7 +276,7 @@ class ProductGrabber(object):
 
             # which shakemap has the highest weight
             weight = 0
-            for idx in xrange(len(eq_info['properties']['products'][sm_str])):
+            for idx in range(len(eq_info['properties']['products'][sm_str])):
                 if eq_info['properties']['products'][sm_str][idx]['preferredWeight'] > weight:
                     weight = eq_info['properties']['products'][sm_str][idx]['preferredWeight']
                     shakemap_json = eq_info['properties']['products'][sm_str][idx]
